Removed custom options.py and Model metaclass
In fact, at first the design of this project is highly influenced by
PirosB3/django-mailer, which is a pioneer project to custom data
backstore(gmail). In that project he customs its own options and
model class, so in this disqus backstore project at first I chose
the same design, too. After thinking I found out that it's not
necessary to make our own Option class for disqus-backstore.
And it's OK to use the django Model as base class of our own class.
What we need to change is the Manager and the Queryset backend.
So in this commit I remove the DisqusBase metaclass,
let the Post and Thread inherit the django.db.models.Model,
and remove the options.py.