Reimplement filter/get with custom Iterable
Before this commit, each time we call `Thread.filter()`, it will
send request to get data immediately and don't cache.
Since in the admin view there are many places calling the filter(),
it costs a lot. Also, the `Post.get()` repeatedly calling the `Thread.get()`
for each post getting from `get_posts_list`, even there's query condition
(e.g primary key). That's a redundant behavior. Also, it's a bad idea to
allow the `data` attribute changed in almost all the Post/Thread methods.

So I think it's a better way to have a same design as Djnago ORM QuerySet:
Use a Iterable class to get data from backstore. Only request data when necessary.
So, now when calling `filter` it will just add the query condition and not really
request the data.
(but, because in the admin we only need to know the 'pk=XXX' query,
so far we haven't(and don't need to) implement a full query parser.)
Also, I try not to request thread data repeatedly in the PostIterable.

Note this change will break the delete operation in admin. This should be fixed.