Fixed mock in test for Post
When Post model instance is created, it needs to find the related
Thread object, so we have to call the `get_thread_list` method to
get thread data and create one. Before this commit I accidentally ignore
the mock for this. This bug cause the very very long test running time,
since each time the post change_list is rendered, it must really send request
to DISQUS api to get the real thread data. Now we add the mock for it,
so all the test can be finished under 1 sec.
Also I replace the way to add mock object. Before this commit we used
the with statement. With the multiple mock objects, it's ugly to have
double with statements. So now we use the `mock.patch.object` decorators
to add mocks.