Added caching in DisqusQuery
Before this commit, the full deletion operation in Admin is very slow,
because in each call of Post/Thread.get it has to send request to DISQUS
to fetch the data. It's about 20 API calls during changelist -> change form
-> delete confirmation -> deletion views and cost 20 sec totally in the process.
So now we cache result when we get data from DISQUS, and clear cache when
we execute update/delete operation.
Note when end user create new post/thread on the forum the cache won't be cleared.
We should have a timer for this to automatically fetch new data if the previous
call is too long before.