Handled DISQUS api Exception
Before this commit we did't check the response from DISQUS. Now we
raise exception if the return code is not zero(e.g. exception occur.)

Not sure if the method coding style in DisqusQuery should be refactored
or not. So far at least, we have a consistent coding style and encapsulate
the requests in a helper function `send_request_to_disqus`.